+++ IIRFactory.h
@@ -6,5 +6,5 @@ namespace ASI
 {
   class InitFilter;
 
-  void createButterBandPassFilter(const int order, const size_t sr, const double lower, const double upper, InitFilter & filter);
+  void createButterBandPassFilter(const size_t order, const size_t sr, const double lower, const double upper, InitFilter & filter);
 }
